Create RemoveIslands
Given a 2-dimensional matrix of 1's and 0's, where 1 represents land and 0 represents water. The objective is to simulate a storm coming by removing all "islands" within the matrix and turn them to "water".  A 1 is considered safe when it is on the border or connected adjacently to one or more ones that are on the border.  A 1 is considered an island when it is not connected adjacently to one or more 1's that lead back to the border.

For example, let's say the matrix inputted was this: 

          [[1, 0, 0, 0, 0, 0],
          [0, 0, 1, 1, 1, 1],
          [0, 1, 0, 1, 0, 0],
          [0, 1, 0, 1, 0, 0],
          [0, 0, 0, 0, 0, 0],
          [1, 0, 1, 0, 1, 0],
          [0, 0, 0, 1, 0, 0]]

The expected output would be:

           [[1, 0, 0, 0, 0, 0], 
           [0, 0, 1, 1, 1, 1],
           [0, 0, 0, 1, 0, 0], 
           [0, 0, 0, 1, 0, 0], 
           [0, 0, 0, 0, 0, 0], 
           [1, 0, 0, 0, 0, 0], 
           [0, 0, 0, 1, 0, 0]]

All of the 1's that were islands are washed away now into the ocean, leaving only the border islands standing.